Reverse Chronological order of Conferences
Displaying the events/conferences in a reverse chronological order
within each year makes it easier to quickly search for which events
are remaining in an year.
This also makes sure that the latest information is available with
minimal scrolling.

Also add relevant instruction about Reverse Chronological order in
the Contributing section.